Basic Progress Saving
Progress is saved to player prefs for each scene as strings of
comma-separated door names that the player has walked through, with the
scene names as the keys. I can't really test it though because I think I
need to update my version of Unity to get rid of errors I'm getting from
the mobile platform particle stuff.